requirements.txt
torch
torchvision
pyquaternion
pyyaml
rospkg
pexpect
mujoco==2.3.7
dm_control==1.0.14
opencv-python
matplotlib
einops
packaging
h5py
ipython
pyserial
keyboard


# cd act/detr && pip install -e .